+++ CHANGES.md
@@ -2,16 +2,17 @@
 
 ## Unreleased
 
+### Added
+
++ core: Add `TypedListView` as idiomatic wrapper over `gtk::ListView`
+
 ### Fixed
 
 + macros: Improve error messages for non-identifier parameter patterns
 
 ## 0.6.0-beta.1 - 2023-4-19
 
-### Added
-
 + core: Introduce setting and action safeties
-+ examples: Introduce setting and action safeties
 + core: Implement `RelmSetChildExt` for `gtk::AspectFrame`
 + core: Add `FactoryHashMap` as alternative to `FactoryVecDeque`
 + core: Add gnome_44 feature flag for GNOME 44
@@ -19,6 +20,7 @@
 + core: Add `set_tooltip` method to `RelmWidgetExt`
 + core: Add `main_adw_application` method to retrieve the `adw::Application` when the libadwaita feature is enabled
 + macros: Add `skip_macro` option for watch and track attributes to skip their initialization 
++ examples: Introduce setting and action safeties
 + examples: Example for using relm4-icons 
 
 ### Changed